requirements.txt
streamlit>=1.38
requests>=2.32
openai>=1.50.0
pandas>=2.0
python-dotenv>=1.0
